getting rid of all unnecesary processes and real data
the aim of this branch is to create a very simple model without the aim to validate the approach via real data. This way we want to prove the concept, and only the concept!

when commiting this branch, there were A LOT of changes listed, especially in venv/Lib/site-packages... I don't know why all these scripts are listed here?